commit message
publish and tag docker images continuously to ghcr.io. adds custom
tagging so we can retain every commit. it's technically possible to do
this on docker hub, but in order to have custom tags, you need to set up
a regular user and log in/push via that user. the authentication is much
cleaner in github actions for ghcr.

(note docker hub pulls for releases are still staying the same, this is
just an alternative form of retention going forward).