chore: fix args passed to validate_call_args
`validate_call_args` takes kwargs, the list of valid keywords as an
argument and makes sure that when a call is made, the given keywords are
valid according to the passed kwargs. however, vyper does not allow
kwargs when calling internal functions, so we should actually pass no
kwargs to `validate_call_args`.

note that this PR does not actually introduce observed changes in
compiler behavior, as the later check in `fetch_call_return` correctly
validates there are no call site kwargs for internal functions.

chainsec june review 5.7