chore: fix loop variable typing
there is an inconsistency between codegen and typechecking types when a
loop iterates over a literal list.

in this code, although it compiles, during typechecking, `i` is typed as
a `uint8` while `[1,2,3]` is annotated with `int8[3]`

```
@external
def foo():
for i in [1,2,3]:
    a: uint8 = i
```

since the iterator type is always correct, this commit is a chore since
it fixes the discrepancy, but there is no known way to "abuse" the
behavior to get a wrong codegen type.

chainsec june 2023 review 5.15