feat: optimize dload/mstore sequences (#3525)
merge_dload was defined in 5dc3ac7, but was not applied.

this commit also adds a rewrite rule for single dload/mstore patterns,
any `(mstore dst (dload src))` (which compiles to a `codecopy` followed
by an `mload` and `mstore`) is rewritten to a `dloadbytes` (which
compiles directly to a `codecopy`).

this rule saves 4 bytes / ~10 gas per rewrite. for instance, it shaves
25 bytes off `VaultV3.vy`, 50 bytes off `CurveTricryptoOptimizedWETH.vy`
and 75 bytes off `CurveStableSwapMetaNG.vy` (basically 0.1%-0.3%).