fix: public constant arrays (#3536)
public getters for arrays would panic at codegen because type
information for the array members was not available. this is because
type annotation would occur before getter expansion. this commit moves
the type annotation phase to right before getter expansion, so that the
generated ast nodes will get annotated.

it also fixes a small bug when trying to deepcopy the nodes generated by
ast expansion - the generated nodes have no node_id and raise an
exception when deepcopy tries to perform `__eq__` between two of the
generated FunctionDefs.