chore: note `Context.in_assertion` is dead (#3564)
the `Context` class has an `in_assertion` flag which, when set, indicates
that the context should be constant according to the definition of
`is_constant()`. however, this flag is never set during code generation,
specifically, it is possible to have a non-constant expression in an
assert statement. for example, the following contract compiles:

```vyper
x: uint256

@internal
def bar() -> uint256:
    self.x = 1
    return self.x

@external
def foo():
    assert self.bar() == 1
```

chainsec june 2023 review 5.5