refactor: initcode generation (#3574)
move internal function generation to after ctor generation. prior to
this commit, the existing code relies on the fact that the code
generation of runtime internal functions properly sets the frame
information of the ctor's callees. if this precondition is not met in
the future, the compiler could panic because the memory allocation info
will not be available.

chainsec june 2023 review 6.2