feat: relax restrictions on internal function signatures
relax the restriction on unique "method ids" for internal methods. the
check used to be there to avoid collisions between external method ids
and internal "method ids" because the calling convention for internal
functions used to involve the method id as part of the signature, but
that is no longer the case. so we can safely allow collision between
internal "method ids" and external method ids.

cf. issue #1687 which was resolved in in 9e8c661494d84fbf.

chainsec june 2023 review 5.22