commit message:
this commit fixes an issue where multiple writes inside of a checkpoint
lead to journal corruption on rollback. in particular, it ensures a call
to `register_update()` when the metadata dict has already been updated
inside of a given checkpoint.

note this does not change any observed functionality in the compiler
because writes to the metadata journal inside for loops only ever happen
to be written once, but it prevents a bug in case we ever add multiple
writes inside of the same checkpoint.

chainsec june review 5.3