fix: `ecrecover()` buffer edge case (#3586)
this commit fixes an edge case in `ecrecover()` that was not covered by
019a37ab98ff5. in the case that one of the arguments to ecrecover writes
to memory location 0, and the signature is invalid, `ecrecover()` could
return the data written by the argument. this commit fixes the issue by
allocating fresh memory for the output buffer (which won't be written to
by evaluating any of the arguments unless the memory allocator is
broken).