fix: version parsing for release candidates (#3593)
the npm spec library is buggy and does not handle release candidates
correctly. switch to the pypa packaging library which does pep440. note
that we do a hack in order to support commonly used npm prefixes: no
prefix, and `^` as prefix. going forward in v0.4.x, we will switch to
pep440 entirely.