chore: reorder compilation of branches in stmt.py
the if and else branches were being allocated out-of-source-order. this
commit switches the order of compilation of the if and else branches to
be in source order. this is a hygienic fix, right now the only thing
that should be affected is the memory allocator (but if more side
effects are ever introduced in codegen, the existing code might compile
the side effects out of order).