fix: metadata output interaction with natspec (#3627)
enabling the `-f metadata` output has an interaction with other outputs
because the metadata output format mutates some internal data structures
in-place. this is because `vars()` returns a reference to the object's
`__dict__` as opposed to a copy of it. the behavior can be seen by
trying to call the compiler with `-f metadata,devdoc,userdoc`. this
issue was revealed in (but not introduced by) 2bdbd846b0, because that
commit caused metadata and userdoc to be bundled together by default.

this commit fixes the issue by constructing a copy of the object during
metadata output formatting. it also modifies the test suite to include
more output formats, to test the interactions between these different
output formats. in doing so, it was also found that some examples have
invalid natspec, which has also been fixed.