chore: remove deadlines and reruns (#3630)
for historical reasons, pytest ran with `--allow-reruns` when tests
failed because tests would usually fail due to deadline errors. deadline
errors have never indicated any meaningful issue with the compiler, they
are just a somewhat unavoidable byproduct of the fact that we are
running in a CI environment which has a lot of jitter.

this commit changes the hypothesis deadline to `None` for the whole test
suite, and removes the `--allow-reruns` parameter in the CI, which
should make the test suite much more efficient when there are failures.