fix: block `mload` merging when src and dst overlap (#3635)
this commit fixes an optimization bug when the target architecture has
the `mcopy` instruction (i.e. `cancun` or later). the bug was introduced
in 5dc3ac7. specifically, the `merge_mload` step can incorrectly merge
`mload`/`mstore` sequences (into `mcopy`) when the source and
destination buffers overlap, and the destination buffer is "ahead of"
(i.e. greater than) the source buffer. this commit fixes the issue by
blocking the optimization in these cases, and adds unit and functional
tests demonstrating the correct behavior.

---------

Co-authored-by: Robert Chen <me@robertchen.cc>