refactor: merge `annotation.py` and `local.py` (#3456)
This commit merges two phases of analysis: typechecking
(`vyper/semantics/analysis/local.py`) and annotation of ast nodes with
types (`vyper/semantics/analysis/annotation.py`). This is both for
consistency with how it is done for `analysis/module.py`, and also
because it increases internal consistency, as some typechecking was
being done in `annotation.py`. It is also easier to maintain this way,
since bugfixes or modifications to typechecking need to only be done in
one place, rather than two passes. Lastly, it also probably improves
performance, because it collapses two passes into one (and calls
`get_*_type_from_node` less often).

This commit also fixes a bug with accessing the iterator when the
iterator is an empty list literal.