feat: search path resolution for cli
the current behavior is that the current directory does *not* get into
the search path when `-p` is specified, which is annoying. (one would
expect `vyper some/directory/some/file.vy` to compile no matter what
`-p` is specified as).

this commit also handles the addition of multiple search paths specified
on the CLI, and adds a long `--path` option as an alternative to `-p`.