refactor: for loop target parsing (#3724)
use a parser trick to get always-correct source locations into for-loop
target annotations.

this fixes an issue with parsing the for loop target (introduced in
ddfce5273b3), where the target annotation is not decorated with the
correct line/column offsets. a6dc432db2df partially fixed the issue, but
does not decorate correctly when the annotation is split across multiple
lines. this commit approaches the problem differently, by using
`untokenize()` to produce a source code for the annotation during AST
massaging which "happens" to have the annotation in exactly the same
place as it appeared in the original source.

small refactors:
- move annotate_python_ast to be earlier in the file. it's somehow
  easier to navigate to it when it comes right before the definition of
  AnnotatingVisitor.
- revert changes from a6dc432db2df to ast/pre_parser.py; go back to
  slurping just the annotation as opposed to the target+annotation
  together.