fix: `opcodes` and `opcodes_runtime` outputs (#3735)
Fixed the opcodes and opcodes_runtime outputs as they would not
respectively match the bytecode and bytecode_runtime outputs.

The value following a PUSH instruction could be incorrect. For example,
when compiling some Vyper code that results in `PUSH2 0x0100`:

- The bytecode output would be `610100`
- The opcodes output would be `PUSH2 0x10` instead of `PUSH2 0x0100`

Note this commit fixes an ambiguity here, as prior to this commit,
`610100` (`PUSH2 0x0100`) and `610010` (`PUSH2 0x0010`) would both get
formatted as `PUSH2 0x10`.

This issue is due to a lack of 0 padding.