perf: levenshtein optimization (#3780)
optimize compile time. `levenshtein` is a hotspot since it is called a
lot during type analysis to construct exceptions (which are then caught
as part of the validation routines). this commit delays calling
`levenshtein` until the last minute, and also adds a mechanism to
`VyperException` so that hints can be constructed lazily in general.

on a couple test contracts, compilation time comes down 7%. however, as
a portion of the time spent in the frontend, compilation time comes down
20-30%. this will become important as projects become larger (that is,
many imports but only some functions are actually present in codegen)
and compilation time is dominated by the frontend.