chore: remove tox rules for lint commands
this commit removes the tox rules for lint, as part of a larger movement
to remove tox from the codebase (it's slow, confusing, and superseded by
environment matrices on github actions).

the rules in the makefile are split apart so they can be run in parallel
with `make -j lint`.