feat[lang]!: change the signature of `block.prevrandao` (#3879)
`block.prevrandao` is an opcode alias for `difficulty`. however, it
returns a bytes object at the evm level (cf. for instance py-evm returns
the value of `block.mixhash`). this commit changes the signature of
`block.prevrandao` to be more consistent with the VM semantics.

also update PR title regex per conventional commit specification - add
optional `!` which indicates a breaking change