fix[test]: fix failure in grammar fuzzing (#3892)
fixes a fuzz failure, which is that the strategy can put utf-8
characters in strings (which we reject at compile-time). the fix here is
to just create programs with ascii. this bug was apparently introduced
in 176e7f7f3a6a16bc09ec58d4bbc8dc515c48a0a8 due to the new alphabet
parameter passed to the grammar strategy.