fix[lang]: pure access analysis (#3895)
this commit fixes the ability to access module variables in pure
functions. the `_validate_self_reference()` utility function was
hard-coded to check the "self" name; remove it and replace with an
analysis-based check.

this commit also fixes the pure access check for immutable variables,
and address members (e.g. `.codesize`)

misc/refactor:
* rename `VarInfo.is_module_variable()` to more fitting
  `VarInfo.is_state_variable()`.
* refactor pure decorator tests to be in line with
  recent best practices