fix[test]: fix a bad bound in decimal fuzzing
the decimal fuzz test `is_valid` condition was based on an ancient
version of decimals which had bounds at -2**127 and 2**127. update the
condition to be compatible with the latest version of Decimal.

also increase the range of decimals produced by the decimal fuzzing
strategy, so that the fuzzer finds overflow issues faster.