fix[lang]: allow type expressions inside pure functions (#3906)
20432c505c706ed introduced a regression where type expressions like the
following would raise a compiler error instead of successfully
compiling:
```
@pure
def f():
    convert(..., uint256)  # raises `not a variable or literal: 'uint256'`
```

the reason is because `get_expr_info` is called on `uint256`, which
is not a regular expr. this commit introduces a fastpath return to
address the issue. longer-term, we should generalize the rules in
`vyper/semantics/analysis/local.py` so that AST traversal does not
progress into type expressions.