feat[lang]!: change ABI type of `decimal` to `int168` (#3696)
this commit changes how decimals are exposed in the ABI - changing
the ABI type from `fixed168x10` to `int168`. this is done for
compatibility reaosns, since `fixedMxN` types are not widely supported.
it is also a preparatory breaking change for supporting more decimal
types, in the future, we might expand the set of decimal types, e.g.
something along the lines of `Decimal[bits, places]`. in such a future,
if we add `UDecimal[256, 18]` (colloquially known as `"wad"`), it would
be useful to users if it had an ABI type of uint256, for compatibility
with common ERCs.

to distinguish from plain `int168` in the JSON ABI, an `"internalType"`
field is added, which includes the decimal info. in the future, if we
add more decimal types, it would be distinguished from other decimal
types according to the metadata, ex.:
`{"internalType": {"decimal": {"bits": 168, "places": 10}}}`.

misc/refactor:
- remove `FixedMxN` abi types
- add a `decimal_to_int()` utility function to make test migration
  cleaner. it essentially emulates what the compiler does internally during
  codegen, which is that it takes the arguments to a `Decimal` and then
  bitcasts the resulting `Decimal` to an int.

---------

Co-authored-by: tserg <8017125+tserg@users.noreply.github.com>