chore[ux]: compute natspec as part of standard pipeline (#3946)
prior to this commit, natspec had to be explicitly requested by the user
(`-f userdoc`, `-f devdoc` or `-f combined_json`). this would lead to
discrepancies between development time and verification time, where a
contract could compile locally (because no natspec was requested), but
not on the verifier's pipeline (because the verifier would request the
natspec via one of the above methods). this commit computes the natspec
as part of the dependencies of the analysed AST (the reasoning being
that, a consumer might expect semantic analysis to include natspec
validation). so, there is no way to produce bytecode for a contract
without validating natspec.